MNIST_UPLOAD/requirements.txt
fastapi==0.115.14
uvicorn[standard]==0.35.0
Pillow==11.2.1
numpy==2.0.2
torch==2.6.0
torchvision
python-multipart